+++ src/pipes/index.ts
@@ -1,3 +1,4 @@
 export * from './displayValue/displayValue.pipe';
 export * from './getPlugin/getPlugin.pipe';
+export * from './groupedListOptions/groupedListOptions.pipe';
 export * from './selectHasValue/selectHasValue.pipe';